Create Digit Transformation.java
Given a string S consisting only of digits (0–9). Your task is to identify all the set (contiguous groups of the same digit) and transform each set based on the following rule: - If the digit is even, replace the entire set with the digit's square mod 10. - If the digit is odd, replace the entire set with the digit's cube mod 10. Return the transformed string.
IP: 112223333
OP: 114447777